Set TextPayload for TCP Access Log and fix source_port showing up as 0 in TCP logs

+++ extensions/common/context.h
@@ -105,10 +105,10 @@ struct RequestInfo {
   int64_t response_size = 0;
 
   // Destination port that the request targets.
-  uint32_t destination_port = 0;
+  uint64_t destination_port = 0;
 
   // Source port of the client.
-  uint32_t source_port = 0;
+  uint64_t source_port = 0;
 
   // Protocol used the request (HTTP/1.1, gRPC, etc).
   std::string request_protocol;
